155 DTHT01 template review (#212)
close #155 

The percentage issue has been temporarily fixed. For more details, see
#211.

The label removing `DEATH DUE TO Death` is fixed by changing the value
in the original dataset, which is similar to the percentage issue. But
it can only detect the issue of `DEATH DUE TO Death` itself. If more
weird cases are needed to correct, then we might try to find a better
way to handle this problem.

Signed-off-by: b_falquet <64274616+BFalquet@users.noreply.github.com>
Co-authored-by: b_falquet <64274616+BFalquet@users.noreply.github.com>
Co-authored-by: benoit <benoit.falquet@roche.com>