add emtpy table tests (#378)
close #357 

during the implementation of these tests, I found something that we need
to be careful

1. arguments for pre, main, and post. they need to be consistent across
functions if the same variable is used, and if, users want to change the
template pre/main/post, it is required that the default arguments are
the same, or not provided. this will ensure that users won't run into
strange issues.
2. arguments for checks/lyt should be exposed in the calling function
(pre and main) so that these arguments are captured correctly in the
`get_argument` functionality.