370 get_extra_args functionality (chevron) (#375)
relates to
https://github.com/insightsengineering/chevron/issues/370
#365 

This is a prototype. The idea is to make the tlg generation process more
transparent for the users to dig in if they want to modify it but do not
which to edit/create a `chevron_tlg` object.

The prototype is based on the following modification:
- the default values and the names of the arguments of each function
stored in a `chevron_tlg` are now saved in the `args` slot of the same
object. If these functions are edited, the `args` slot is also
automatically modified to keep consistancy.
- the `args` method allows to access the `@args` slot value. It can
provide the arguments for each of the pre/main/post function separately
or in a simplified version (when `simplify = TRUE`).
- the `script` method creates a script based on a chevron_tlg object

See the newly added `Script Generator` vignette for more details

Question ----
- For now the script generates the entire workflow (it prints the
preprocess, main and postprocess function together with some additional
assignment). Should we limit it to just the preprocessing ?

thank you for the review.

NOTE: step to finalize the PR ----

- [x] add more tests for `args`
- [x] add more tests for `script`
- [x] improve tests for `preprocess`/`main`/`postprocess` to take into
account the modification of `@args`
- [x] add to pkgdown

---------

Co-authored-by: benoit <benoit.falquet@roche.com>
Co-authored-by: Liming <36079400+clarkliming@users.noreply.github.com>