correct failure due to timezone handling (#750)
correct error in integration tests.

In the listings, the time reported should be the time of the event at
the location where the event was recorded. Using base function changes
the result depending on the time zone of the computer. Use `lubridate`
to circumvent the issue.


thank you for the review.