Replace logo.png by logo.svg (#800)
# Pull Request

Replace the logo.png by logo.svg to reduce package's size (This is
relevant for CRAN submissions were packages cannot be generally >5Mb of
source code and get flagged for CRAN reviewer if they exceed this
number).

Important: the logo appears to be different: there is some green pattern
on the SVG that is not on other logo formats.

This package makes use of Github actions to check the branding, replaced
there too.

Fixes https://github.com/insightsengineering/nestdevs-tasks/issues/93